fix migration on mysql
only postgres can use the conditions, and mysql complains about us removing
the index supporting the FK, so only do it for postgres

Reviewed-on: https://gerrit.instructure.com/24386
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>

+++ db/migrate/20130820202205_add_index_on_course_wiki_id.rb
@@ -4,13 +4,17 @@ class AddIndexOnCourseWikiId < ActiveRecord::Migration
 
   def self.up
     add_index :courses, :wiki_id, concurrently: true, conditions: "wiki_id IS NOT NULL"
-    remove_index :groups, :wiki_id
-    add_index :groups, :wiki_id, concurrently: true, conditions: "wiki_id IS NOT NULL"
+    if connection.adapter_name == 'PostgreSQL'
+      remove_index :groups, :wiki_id
+      add_index :groups, :wiki_id, concurrently: true, conditions: "wiki_id IS NOT NULL"
+    end
   end
 
   def self.down
     remove_index :courses, :wiki_id
-    remove_index :groups, :wiki_id
-    add_index :groups, :wiki_id
+    if connection.adapter_name == 'PostgreSQL'
+      remove_index :groups, :wiki_id
+      add_index :groups, :wiki_id
+    end
   end
 end
